Added control of slider values when variable is only round numbers
isallinteger - added function to check whether all values in a vector are round numbers.
setslider - add call to isallinteger to set mid-point of slider and force duration and round number doubles to integer values for the slider control (so that only round numbers can be selected)
inputUI - add call to isallinteger when resetting a slider UI. force rounding when updating variables that are round numbers (including durations).
get_precision - added check that input is a double and return p=0 if not.