Improve sort using index workflow and add check when loading an App file into an App
muiModelUI - add trap to prevent user loading wrong mat file into an App
muiTableImport - improve workflow when sorting data using an index
muiDataSet - return index as well as dataset name in getDataSetName
sort_var -input changed to class instance and dataset index. options to prompt user to select variable or pass in variable index or the variable itself as a vector. (NB selection is for a variable only but can pass in a dimension to sort).