Fix data selection UIs that subsample
NB: important fix. The data selection UIs that subsample were not updating if the user changed the order of the variables in the subselection. These changes fix this issue:
muiDataUI - pass order of variables to inputUI and return the selected order
inputUI - added InputOrder field to input struct and UIorder to output so that any change in variable order is tracked. Modified handling of durations so that a duration is returned when single value selected from slider (slider does not accept durations)
setslider - add UserData to slider value text so that durations can be tracked
editrange_ui - add empty InputOrder field for consistency with inputUI
muiCatalogue - assign default row name in getSelectedIndices and handle integers when interpolating selection in getIndices.