Adjust save to handle different files sizes, improve plotting or roses and function for binning timeseries data
Added trap in muiModelUI.saveModel to try using -v7 and if this fails (for large data sets where sobj >2^31 bytes) use -v7.3.
Added function to bin a timeseries variable by intervals such as months or years.
Extended plotting of Roses in muiPlots to allow user to control the intervals for intensity and %
Improved documentation in clusters and getclusters
In wind_rose replace code to components to top of UI stack to use uistack